HSD #15016025158-2: ddr: altera: iossm: Mailbox cmd error handling refactor
When an error happens during IOSSM mailbox command running, it continues
to execute the subsequent commands, causing new issues. This makes it
misleading to analyze the original mailbox command error. Instead of
continuing to execute the following commands, we would trigger a U-Boot
hang when any mailbox error happens. This way, the driver will stop at
the error location and not proceed to the subsequent code. This would
help reduce confusion when analyzing mailbox command-related issues.

Signed-off-by: Tingting Meng <tingting.meng@intel.com>