HSD #15016529058: arm: socfpga: agilex5: Get sub device type info
This patch enables the workaround by checking the device type at runtime
and applying the necessary fix for the effected version. The JTAG ID is
read from the boot scratch cold 4 register, and based on the sub device
type info retrieved from JTAG ID, the appropriate code path is executed
to correct the erratum on the silicon.

Signed-off-by: Tingting Meng <tingting.meng@intel.com>