HSD #15016500042: drivers: clk: agilex5: Configure intosc as boot_clk source
This patch enables the configuration of the HPS internal oscillator as
the boot_clk source instead of the default external
oscillator(HPS_OSC_CLK) in non-secure boot scenarios. Some customers
prefer to minimize the use of external oscillators, especially when using
the FPGA first configuration mode. This change allows them to eliminate
the need for an additional oscillator device and a dedicated HPS pin,
simplifying board layout and routing.

Signed-off-by: Tingting Meng <tingting.meng@intel.com>