HSD #14023071488: ddr: altera: agilex7m: Update calibration status checking method
The Calibration Busy status is now set to low by default. In the previous
design, U-Boot relied on the Calibration Busy status to determine whether
DDR calibration was complete. However, if U-Boot polling the calibration
status before calibration is started, it could mistakenly interpret the
default low value of the Calibration Busy bit as indicating that
calibration is completed. This led to U-Boot reading the invalid
calibration result and resulted to system hang for waiting recovery. To
address this issue, we now poll the calibration success and fail status
directly instead of relying on the Calibration Busy status.

Signed-off-by: Tingting Meng <tingting.meng@intel.com>