HSD #14023897270: arch: arm: mach-socfpga: Notify SDM of FSBL boot execution.
Call mbox_hps_stage_notify() in SPL to inform the
SDM of the first stage bootloader execution. This
notification is necessary for the SDM to recognize
that the FSBL has begun its execution and should be
made as early as possible in the FSBL process.
Ensure that mbox_init() is called after timer_init().

Signed-off-by: Mahesh Rao <mahesh.rao@intel.com>