configs: agilex5: Remove duplicated config
Remove duplicated bloblist address config in Agilex5 defconfig.

This bloblist address is used in upstreamed Agilex5 defconfig,
in which the SMMU translation table was not included in the SPL
stack, resulting in different fixed bloblist address after
calculation.

SMMU workoaround is used for Agilex5 A0 devkits. We should use
the upstreamed address once we decided to no longer support A0.

Signed-off-by: Alif Zakuan Yuslaimi <alif.zakuan.yuslaimi@altera.com>