HSD #15017844483: arch: arm: dts: Increase JFFS2 RootFS partition size
The current partition size for the JFFS2 RootFS in the NAND device tree
(0x05500000) is too small for the actual filesystem size. This is
causing mount failures when the filesystem exceeds the partition boundary.
The partition size increased to 0x3db00000 to ensure sufficient space
for the JFFS2 RootFS and prevent filesystem errors.

Signed-off-by: Dinesh Maniyam <dinesh.maniyam@altera.com>