HSD #15017984905: arm: socfpga: Agilex5 backwards compatibility fix
There is an issue with I3C0 and I3C1 section of FPGA use section of the
handoff data prior to Quartus 25.1.1 builds which causes memory violation
errors as the incorrect registers were to be accessed.

This commit will try to read the offset values of I3C0 and I3C1 from the
handoff first and will then proceed to write to the registers after
verifing the correct offset values has been  received from the handoff
data.

This is a temporary workaround to support backwards compatibility of older
Quartus builds, this commit should not be upstreamed and will be dropped
after several stable releases.

Signed-off-by: Alif Zakuan Yuslaimi <alif.zakuan.yuslaimi@altera.com>